Primeiro Commit no Branch  DNX. Limpeza geral no código, e separação das dependências de WebSites na classe principal.
Funcionalidades que usam controls e classe Page retiradas. Poderão ser reincorporadas futuramente.

Reengenharia nas classes básicas, todas as funcionalidades específicas dos bancos foram retiradas. Uma Pasta Common foi criada para abrigar as classes bases dos Bancos em separado.
