+++ src/.vs/config/applicationhost.config
@@ -163,10 +163,18 @@
             </site>
             <site name="Boleto.Net.Site" id="2">
                 <application path="/" applicationPool="Clr4IntegratedAppPool">
-                    <virtualDirectory path="/" physicalPath="D:\cds\dotnet\OpenSource\boletonet\boletonet\src\Boleto.Net.Site" />
+                    <virtualDirectory path="/" physicalPath="D:\Dev\BoletoNet\src\Boleto.Net.Site" />
                 </application>
                 <bindings>
                     <binding protocol="http" bindingInformation="*:3304:localhost" />
+                </bindings>
+            </site>
+            <site name="Boleto.Net.MVC" id="3">
+                <application path="/" applicationPool="Clr4IntegratedAppPool">
+                    <virtualDirectory path="/" physicalPath="D:\Dev\BoletoNet\src\Boleto.Net.MVC" />
+                </application>
+                <bindings>
+                    <binding protocol="http" bindingInformation="*:26014:localhost" />
                 </bindings>
             </site>
             <siteDefaults>
