+++ src/Boleto.Net.MVC/App_Start/WebApiConfig.cs
@@ -1,19 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Web.Http;
 
 namespace Boleto.Net.MVC
 {
     public static class WebApiConfig
     {
-        public static void Register(HttpConfiguration config)
-        {
-            config.Routes.MapHttpRoute(
-                name: "DefaultApi",
-                routeTemplate: "api/{controller}/{id}",
-                defaults: new { id = RouteParameter.Optional }
-            );
-        }
+        
     }
 }
