+++ src/Boleto.Net.MVC/packages.config
@@ -29,5 +29,6 @@
   <package id="Modernizr" version="2.5.3" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="4.5.6" targetFramework="net45" />
   <package id="NReco.PdfGenerator" version="1.1.7.0" targetFramework="net40" />
+  <package id="System.Net.Http" version="2.0.20710.0" targetFramework="net40" />
   <package id="WebGrease" version="1.1.0" targetFramework="net45" />
 </packages>
