+++ src/Boleto.Net.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+# Visual Studio 14
+VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Biblioteca", "Biblioteca", "{75FC8289-5B1E-4684-8497-54494382FA55}"
 	ProjectSection(SolutionItems) = preProject
@@ -30,32 +30,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Boleto.Net", "Boleto.Net\Boleto.Net.csproj", "{F173E80B-E36B-48AD-B483-A27F8CA86F3C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Boleto.Net.Arquivo", "Boleto.Net.Arquivo\Boleto.Net.Arquivo.csproj", "{98AD1B9D-47F5-432C-85BD-49A12FBD7CAD}"
-EndProject
-Project("{E24C65DC-7377-472B-9ABA-BC803B73C61A}") = "Boleto.Net.Site", "Boleto.Net.Site\", "{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}"
-	ProjectSection(WebsiteProperties) = preProject
-		SccProjectName = "SAK"
-		SccAuxPath = "SAK"
-		SccLocalPath = "SAK"
-		SccProvider = "SAK"
-		TargetFrameworkMoniker = ".NETFramework,Version%3Dv4.5.1"
-		ProjectReferences = "{F173E80B-E36B-48AD-B483-A27F8CA86F3C}|Boleto.Net.dll;"
-		Debug.AspNetCompiler.VirtualPath = "/Boleto.Net.Site"
-		Debug.AspNetCompiler.PhysicalPath = "Boleto.Net.Site\"
-		Debug.AspNetCompiler.TargetPath = "PrecompiledWeb\Boleto.Net.Site\"
-		Debug.AspNetCompiler.Updateable = "true"
-		Debug.AspNetCompiler.ForceOverwrite = "true"
-		Debug.AspNetCompiler.FixedNames = "false"
-		Debug.AspNetCompiler.Debug = "True"
-		Release.AspNetCompiler.VirtualPath = "/Boleto.Net.Site"
-		Release.AspNetCompiler.PhysicalPath = "Boleto.Net.Site\"
-		Release.AspNetCompiler.TargetPath = "PrecompiledWeb\Boleto.Net.Site\"
-		Release.AspNetCompiler.Updateable = "true"
-		Release.AspNetCompiler.ForceOverwrite = "true"
-		Release.AspNetCompiler.FixedNames = "false"
-		Release.AspNetCompiler.Debug = "False"
-		VWDPort = "3304"
-		SlnRelativePath = "Boleto.Net.Site\"
-	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Boleto.Net.Testes", "Boleto.Net.Testes\Boleto.Net.Testes.csproj", "{92951948-996F-4F22-AB35-74B06C45E1D4}"
 EndProject
@@ -91,16 +65,6 @@ Global
 		{98AD1B9D-47F5-432C-85BD-49A12FBD7CAD}.Release|Any CPU.Build.0 = Release|Any CPU
 		{98AD1B9D-47F5-432C-85BD-49A12FBD7CAD}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{98AD1B9D-47F5-432C-85BD-49A12FBD7CAD}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}.Debug|.NET.ActiveCfg = Debug|Any CPU
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}.Release|.NET.ActiveCfg = Debug|Any CPU
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}.Release|Any CPU.ActiveCfg = Debug|Any CPU
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}.Release|Any CPU.Build.0 = Debug|Any CPU
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}.Release|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167}.Release|Mixed Platforms.Build.0 = Debug|Any CPU
 		{92951948-996F-4F22-AB35-74B06C45E1D4}.Debug|.NET.ActiveCfg = Debug|Any CPU
 		{92951948-996F-4F22-AB35-74B06C45E1D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{92951948-996F-4F22-AB35-74B06C45E1D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
@@ -130,7 +94,6 @@ Global
 		{057F8124-1BE5-46CB-B2EF-106CD4325B93} = {093054D8-8738-4790-87CC-25378443A115}
 		{F173E80B-E36B-48AD-B483-A27F8CA86F3C} = {75FC8289-5B1E-4684-8497-54494382FA55}
 		{98AD1B9D-47F5-432C-85BD-49A12FBD7CAD} = {75C4F877-9574-43BA-9844-EAEC221DC0FA}
-		{3E75E02F-A49C-4C96-8A9E-8A405D8B6167} = {057F8124-1BE5-46CB-B2EF-106CD4325B93}
 		{92951948-996F-4F22-AB35-74B06C45E1D4} = {8C6CE2CA-61F8-489B-A803-27437835E300}
 		{ED82A7C4-89F9-4481-ADE5-8B5F64DF9E5B} = {717CBD6F-C429-445D-95B0-191CEAA7A088}
 	EndGlobalSection
