+++ src/Boleto.Net/Banco/Banco_BRB.cs
@@ -1,8 +1,6 @@
 using System;
-using System.Web.UI;
 using BoletoNet.Util;
 
-[assembly: WebResource("BoletoNet.Imagens.070.jpg", "image/jpg")]
 namespace BoletoNet
 {
     /// <author>
