+++ src/Boleto.Net/Banco/Banco_Basa.cs
@@ -1,9 +1,6 @@
 using System;
-using System.Web.UI;
 using BoletoNet.Util;
 using System.Text;
-
-[assembly: WebResource("BoletoNet.Imagens.003.jpg", "image/jpg")]
 
 namespace BoletoNet
 {
