+++ src/Boleto.Net/Banco/Banco_Nordeste.cs
@@ -2,13 +2,10 @@
 using System;
 using System.Data;
 using System.Globalization;
-using System.Web.UI;
-using Microsoft.VisualBasic;
 using BoletoNet.EDI.Banco;
 using System.Collections.Generic;
 using BoletoNet.Util;
 
-[assembly: WebResource("BoletoNet.Imagens.004.jpg", "image/jpg")]
 namespace BoletoNet
 {
     /// <summary>
